Added _ACRONYM property types, introduced a static property cache
_ACRONYM: Changes the expected getter/setter names (property name "htmlContent" => getter name "getHTMLContent" instead of "getHtmlContent")

Static property cache: a static var $_properties contains a list of property sets assigned to a class name. The first time a class is instanced, the property set for this class is created and it is checked whether all getter and setter methods exist. The following times the same class is instanced, the set of properties is read out of the property cache. A var $_prop (reference to $_properties[class_name]) stores the property set of an instance.