Added missing use statements, removed $index porperties and getters of models and corrected some bugs
Some use statements of the Premanager\QueryList were missing, some $_descriptor and $_queryList static properties too.

The property $index of a model contained the position of the model in "default order". This will be soonly replaced by the indexOf() method of a QueryList, wich will support order() so there will no longer be a "default order".

Some constructors of the models were private instead of protected, and the "static" keyword was missing at some methods .