Ensured that accessing the defualt Premanager\Execution\Environment does not cause a recursion
Former, when accessed the first time an option, the Options class loaded user-defined options-too. When the current user was not detected already, the environment class checked the session cookie to find out who is logged in. A call of Request::getCookie would need the cookie prefix option. At this point, the options would still not be loaded, so the Options class would get the user-defined options, the environment class would try to detect the logged-in user... - we've got an infinite loop.

Now, the Environment class remembers if a property is loading at the moment, and returns a default value if it is accessed the second time. The Options class does not load the user-defined options if the $me property is not available (Environment::isMeAvailable). They will be loaded at the next option access, when the $me property is available.

The only resoluting problem is that the cookie prefix can not be overwritten by a user. But of course, this is technically impossible, just as well as expeting any option to be overwritten by user if the user is not yet detected.

The same problem occurs with $project and project-defined option values, and the same changes have be done to it, as well as the three other properties $language, $theme and $edition.