Premanager\Execution\Template provides access to the Dwoo template engine
The Template class is linked to a template file in the "template" folder of a plugin with the file extension ".tpl".

The set() method can be used to assign data to the collection and clear() to remove all those data. get() renders the template using the provided data and returns the result.

By default, the output is compressed. This means that all unnecessary whitespaces are removed to make the file, e.g. a html or css file, as small as possible. This feature can be disabled for a specific template and is turned off completely in debug mode.