The PageNode::getPage() method gets the page associated with the page node
The getPage() method should always return the page that would be sent to output when calling execute(). The best implementation of execute() is to call Output::select($this->getPage()), but only if you are sure that the page node can only contain a page (and not binary data for example).

Premanager\Execution\StructurePageNode calls the execute() method of the tree class if exists, else selects the page which is a list of subpages or a panel (the structure node's type decides). The panel and list creation is not yet implemented.