Added PageNode::fromPath which finds a page node specified by its url
The method first requests the page node for the root structure node of the current project, then walks through the path parts of the url and calls getChildByName() of each node to get its child. On failure, stores the last found node in $impact, for example to redirect to it.

The abstract method getChildByName() is implemented by StructurePageNode which calls either the same method of the embedded tree node or detects the structure node and then creates a StructurePageNode that represents the structure node.