Fixed Types::isInteger()
Former, isInteger() returned true to all strings that started with digits, like "12abc" or "2010-17-09". This happened because the value was converted into an integer and then compared to the original value. But on comparing ints and strings, the string is converted into an integer - and the expression will always be true.

Now, isInteger() converts the int-converted string into a string before comparing it to the original string.