The Output class can output instances of Page
Use the select() function if you want to output a page. You can specify if the resource have been found and whether it can be accessed by the $statusCode parameter.

You have to call the finish() method to output the page. But this is the job for the Premanager class.