TREE StructurePageNodes are replaced by the page nodes of the structure nodes
If a structure node contains a TREE node and the page node of that TREE node is requested, formerly a StructurePageNode was created which redirected most of the methods to its embedded tree page node.

Now, the page node of a TREE node is the direct child of its parent structure node.

So tree page nodes do no longer have to extend the TreePageNode or TreeListPageNode classes but are common page nodes except they expect a structure node as parameter for the structure which allows them to redirect the name and title requests to the structure nodes.