Security feature: post data are only accepted if they contain the correct session key
If you are logged on as an administrator, a hacker simply has to execute a POST request to do whatever he wants, because the session key is sent in a cookie.

The new security feature rejects post data if they do not contain the session key. This makes sure that the request is done with a form from a Premanager page or with javascript from the Premanager site, because all others do not have access to the session key.