Removed the Premanager\Execution\PageNode::getPage() method
As noticed in the LoginPage, the implementation of two separate methods (getPage and execute) would be very difficult for some page nodes. Their behaviour depends on the POST data, so there are several possibilities for what the page node outputs. For example, a login form, a login-failed notice or a redirection. Additionally, the LoginPage has creates a new session if the login data is correct and this would not be expected a getPage() method to do.