Renamed the "me" property of the environment to "user"
In earlier versions of Premanager, the logged-in user was called "me" and this name was not changed at first. I think that it is not clear what "me" menas, but "user" is clear because the environment contains only one user.

Notice: Other usages of "me" like in translation names are not changed.