Optimized Premanager\Execution\ListPageNode
Former, the count of items was requested in the constructur (they are needed to calculate current page and page count). So the items were counted even if the page properties would not be requested on this instance.

Now, the methods that need the count calculate their value when they are called the first time.