Replaced access to properties by calls of the getter methods for performance reasons
When a property is accessed using the property style ($object->property), this will take a lot of time. First, the __get() method is called which checks if the property is defined and readable and lookups the getter function name. Then, it calls that getter using call_user_func. This is a lot more time-consuming than calling the getter directly.

Currently, the property style is still used at some places, for example in Premanager\QueryList\MemberInfo::getValue() and in the templates. When this has changed, we can remove properties completely which would save time again. The constructor of Premanager\Module (which only prepares the properties) requires 3% of the total page load count of the home page.

The properties where replaced by the getters with an regular expression replacement which could not change the first letter of the property name to upper case. For example, $obj->value got $obj->getvalue() instead of $obj->getValue(). This should be corrected gradually in the next time.