Revised the nav-content-sidebar flow layout of the common style
This is the first working implementation of an important design idea. The navigation tree should be at the left, the sidebar at the right and the content in the middle. The whole thing should be as high as the highest part. This is important, becasue the footer should be below and full-wide.

The former implementation used the CSS "position: absolute" property. It wasn't nice, and I knew that, but it worked. When I introduced the footer, I noticed that the quick-and-dirty-way did no longer work.

Now both navigation tree and sidebar are placed in HTML before the content element. CSS tells them to float around the content. The content element has at the left and right a padding as wide as the both side bars which limits the width of the blocks to the actual available space.