Changed the page node execution system: Introducted the class Premanager\Execution\Response and the method Premanager\Execution\PageNode::getResponse()
The Response class is an abstract class for all objects that can be sent to the client (using Output::send()). They provide the status code, content type and the content itself. For example, Page or Redirection are inheriting classes.

The getResponse() method replaces execute() and returns the response instead of directly sending it to the client. That brings the advantage that the response can be analyzed before sending it. When a page is accessed via AJAX, pages should be send in a different way (e.g. the sidebar should not be included).