Introduced the "backend pages", a way AJAX applications can access pages
The "backend pages" normally result in xml documents (it might be possible to output binary data for performance reasons). This is very useful for JavaScript applications that need to communicate with the server.

A "backend page" is accessed with the following URL: !<PluginName><Path> (notice the exclamation mark). <PluginName> is the qualified name of a plugin that supports backend pages, and <Path> is any path and / or query string that can be evaluated by the PHP script of the backend page.

Every plugin can have a backend tree class, a class extending PageNode. In the page tree it is placed as a child of the root node (Organization). Every backend tree class can have children who are accessed using the getChildByName() method.

Notice: There are a lot of changes in the premanager.sql file. Most of them poorly are results of the clock change of my local machine from CEST (UTC+02:00) to CET (UTC+01:00)