Implemented the smart pageload
Smart pageload is an AJAX technology which speeds the page load up by only loading the real content of the page. This data is requested and received by java script and inserted into the DOM tree.

In future, I will use the location api of HTML5 in order to update the url during using smart pageload.