Introduced the Premanager\QueryList\QueryListStrategy feature
QueryListStrategy is an abstract class that provides all methods that are needed for the QueryList class. QueryList finds the best strategy itself according to the input data

Now, two implementations are available: FullQuery and PHP. FullQuery can only be used if the filter and all sort rules can be represented in a query string and therefore is very quickly because it does not need any post-query operations. PHPQueryListStrategy contains the algorythm QueryList had used before. It can be used as alternative if the filter or a sort rule needs post-query evaluation.