Optimized the Premanager\QueryList\QueryListIterator
An iterator needs a method that checks whether an index is valid. This formerly was done by comparing it to the count of items, even if the count has not been received yet. So another query was needed.

Now, QueryListStrategy has a method called isIndexValid() that does this job. The basic implementation does the same as before - comapring the parameter to the count which is got by getCount() - but the FullQueryQueryListStrategy has an optimized implementation which either does not need any additionally query or combines it with receiving another item.