Added the property pageNode to Premanager\Execution\Environment and included the session property into the constructore ::create()
The pageNode refers to the requeted page node. Normally, it equals Premanager\IO\Request::getPageNode(). But for example the smart-pageload feature can modify this environment variable to pretend as if another page was accessed.

The dwoo form plugin which creates the <form> elements now uses this page node instead of the one of Request.