Improved the time measuring system
All data base connections submit their query count and query time to the Premanager\IO\DataBase\DataBase class which collects it and provides the sum of all.

Smart-pageload now shows the time info like normal page access does to, but adds the time that is needed additionally (for network and DOM tree buliding).