Added Premanager\Execution\FormPageNode
This class allows to create form pages very easily. It only nedds to know which template to use, how to read the form values from the model and from POST data, whether they are valid and what to do with them when the user submits the form and all data is vaild.

The dwoo plugin formElement can create a form element in a simple way: There are only parameters for the field name, a label, an optional description and the input type. It automatically inserts the value and shows errors.