Introduced the "tree class scopes"
The scope of a tree class indicates whether it can be assigned to the organization project to other projects or to all projects.

For example, the pages related to administration and members are only for the organization, project information is only for projects and a blog may be available in both projects and organization.