Introduced the page toolbars
Premanager\Execution\Page is extended by the property $toolbar, an array of Premanaer\Execution\ToolBarItem's. Items can be created using the constructor which expects at least the target url and a title, but can also include a description, url to icon or disable the button.

Because the icons are now linked directly, they are moved from the style directory into the images folder. They belong to no style.