Moved the $model propety of Premanager\Execution\FormPageNode into the concrete classes (i.e. Premanager\Pages\EditProjectPage)
A form page has nothing to do with the model it can edit, because the FormPageNode only needs to know the initial values, how to load and validate the POST data and what to do when validation succeeds. Deciding whether to use a model or not is better off in the conrecte classes.

Additionally, added the EditProjectPage.