Introduced tree class keys and added Premanager\Execution\PageNode::getTreeURL()
The tree class keys are short strings that identify a tree class within plugin context. Formerly, class names could be used for that, but now they are very long because they include the namespace information, and in addition, the class names may change.

getTreeURL() gets the url to a page node that contains a specified tree class.