Introduced style.xml and added the new style "Yogularm"
Instead of a class that calls methods for adding stylesheets, the stylesheets are indexed in an xml file which is placed in the static folder. Later, they can be easily loaded via JavaScript.

The style "Yogularm" is based on the style of http://www.yogularm.de/. The navigation forms a curved bar at the top and left side. The content styles are very light. The style does not use any graphics for the layout, but CSS3 features. Full functionality currently is only available in Firefox.