Sidebar Admin page
The default sidebar now can be edited in a page that provides all available widget classes each with a button to add it to the sidebar. When viewing that sidebar admin page, the default sidebar is displayed instead of the personal one and all widgets have buttons to remove them or move them up and donw.

Only tested with classic style.